+++ ocean_provider/routes/compute.py
@@ -14,8 +14,8 @@ from flask_sieve import validate
 from ocean_provider.requests_session import get_requests_session
 from ocean_provider.user_nonce import update_nonce
 from ocean_provider.utils.basics import LocalFileAdapter, get_provider_wallet, get_web3
+from ocean_provider.utils.compute_environments import get_c2d_environments
 from ocean_provider.utils.error_responses import error_response
-from ocean_provider.utils.provider_fees import get_c2d_environments
 from ocean_provider.utils.util import (
     build_download_response,
     get_compute_endpoint,
