+++ ocean_provider/routes/consume.py
@@ -3,20 +3,20 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 import logging
-import os
 
 from flask import jsonify, request
 from flask_sieve import validate
 from ocean_provider.myapp import app
 from ocean_provider.requests_session import get_requests_session
 from ocean_provider.user_nonce import get_nonce, update_nonce
+from ocean_provider.utils.asset import get_asset_from_metadatastore
 from ocean_provider.utils.basics import (
     LocalFileAdapter,
-    get_asset_from_metadatastore,
     get_provider_wallet,
     get_web3,
     validate_timestamp,
 )
+from ocean_provider.utils.compute_environments import check_environment_exists
 from ocean_provider.utils.error_responses import error_response
 from ocean_provider.utils.provider_fees import get_provider_fees, get_c2d_environments
 from ocean_provider.utils.services import ServiceType
@@ -24,7 +24,6 @@ from ocean_provider.utils.url import append_userdata, check_url_details
 from ocean_provider.utils.util import (
     build_download_response,
     check_asset_consumable,
-    check_environment_exists,
     get_compute_info,
     get_download_url,
     get_metadata_url,
