+++ ocean_provider/utils/basics.py
@@ -4,7 +4,6 @@
 #
 from datetime import datetime
 import os
-from pathlib import Path
 from typing import Optional, Union
 
 import artifacts
@@ -13,7 +12,6 @@ from eth_account import Account
 from hexbytes import HexBytes
 from ocean_provider.config import Config
 from ocean_provider.http_provider import CustomHTTPProvider
-from ocean_provider.utils.asset import Asset
 from requests_testadapter import Resp
 from web3 import WebsocketProvider
 from web3.main import Web3
@@ -110,16 +108,6 @@ class LocalFileAdapter(requests.adapters.HTTPAdapter):
     ):
 
         return self.build_response_from_file(request)
-
-
-def get_asset_from_metadatastore(metadata_url, document_id):
-    """
-    :return: `Ddo` instance
-    """
-    url = f"{metadata_url}/api/aquarius/assets/ddo/{document_id}"
-    response = requests.get(url)
-
-    return Asset(response.json()) if response.status_code == 200 else None
 
 
 def send_ether(web3, from_wallet: Account, to_address: str, amount: int):
